Making Kernel object oriented
